captcha: accept hash with quotes too.
The captcha asks the user to run a nix-instantiate command.
The output of this command is
```
$ nix-instantiate --eval --expr 'builtins.hashString "sha256" "NixOS wiki"'
"62e65110a5a6fa4f08256f7d9ee3461412babb37dc0955531b79dcf9732c9c91"
```

Which is *not* the accepted answer, since the accepted answer does not include the quotes surrounding the hash.

Fix this by accepting the version with quotes too.

I have not tested this, but per the QuestyCaptcha docs, it should work.